chore(routing): restructure routing
