+++ src/lib/config/routes.ts
@@ -1,4 +1,4 @@
 export const routes = {
 	home: "/",
-	new: "/r/new",
+	new: "/new",
 };
