--- src/routes/r/new/+page.svelte
@@ -1,7 +1,0 @@
-<script lang="ts">
-import Radar from "$lib/components/radar/app/organisms/radar.svelte";
-</script>
-
-<main class="flex flex-1 flex-col items-center justify-center">
-  <Radar />
-</main>
